+++ simulator-core/src/main/java/com/evbox/everon/ocpp/simulator/station/actions/Plug.java
@@ -43,23 +43,35 @@ public class Plug implements UserMessage {
 
         evse.plug(connectorId);
 
-        stationMessageSender.sendStatusNotificationAndSubscribe(evse, evse.findConnector(connectorId), (statusNotificationRequest, statusNotificationResponse) -> {
-            if (evse.hasTokenId()) {
-                String tokenId = evse.getTokenId();
-                log.info("Station has authorised token {}", tokenId);
+        if (stationState.isAwaitingForPlug(evseId)) {
+            stationState.stopConnectionTimeOut(evseId);
+            startCharging(evse);
 
-                stationMessageSender.sendTransactionEventUpdateAndSubscribe(evse.getId(), connectorId, TransactionEventRequest.TriggerReason.CABLE_PLUGGED_IN, tokenId, TransactionData.ChargingState.EV_DETECTED,
-                        (transactionEventRequest, transactionEventResponse) -> {
-                            evse.lockPluggedConnector();
-                            evse.startCharging();
+            stationMessageSender.sendTransactionEventUpdate(evseId, connectorId, TransactionEventRequest.TriggerReason.CHARGING_STATE_CHANGED, TransactionData.ChargingState.CHARGING);
+        } else {
 
-                            stationMessageSender.sendTransactionEventUpdate(evse.getId(), connectorId, TransactionEventRequest.TriggerReason.CHARGING_STATE_CHANGED, tokenId,
-                                    TransactionData.ChargingState.CHARGING);
-                        });
-            } else {
-                stationMessageSender.sendTransactionEventStart(evse.getId(), connectorId, TransactionEventRequest.TriggerReason.CABLE_PLUGGED_IN, TransactionData.ChargingState.EV_DETECTED);
-            }
-        });
+            stationMessageSender.sendStatusNotificationAndSubscribe(evse, evse.findConnector(connectorId), (statusNotificationRequest, statusNotificationResponse) -> {
+                if (evse.hasTokenId()) {
+                    String tokenId = evse.getTokenId();
+                    log.info("Station has authorised token {}", tokenId);
+
+                    stationMessageSender.sendTransactionEventUpdateAndSubscribe(evseId, connectorId, TransactionEventRequest.TriggerReason.CABLE_PLUGGED_IN, tokenId, TransactionData.ChargingState.EV_DETECTED,
+                            (transactionEventRequest, transactionEventResponse) -> {
+                                startCharging(evse);
+
+                                stationMessageSender.sendTransactionEventUpdate(evseId, connectorId, TransactionEventRequest.TriggerReason.CHARGING_STATE_CHANGED, tokenId,
+                                        TransactionData.ChargingState.CHARGING);
+                            });
+                } else {
+                    stationMessageSender.sendTransactionEventStart(evseId, connectorId, TransactionEventRequest.TriggerReason.CABLE_PLUGGED_IN, TransactionData.ChargingState.EV_DETECTED);
+                }
+            });
+        }
+    }
+
+    private void startCharging(Evse evse) {
+        evse.lockPluggedConnector();
+        evse.startCharging();
     }
 
 }