+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Station Simulator Releases #
 
+### Version 0.3.5
+
+- Added support for Remote start transactions
+- Added support for Remote stop transactions
+- Small improvements
+
 ### Version 0.3.4
 
 - Improvements to the console tools